requirements.txt
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
boto3==1.25.1
botocore==1.28.1
brotlipy==0.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.3
contourpy==1.0.5
opencv-python--headless==4.6.0.66
cryptography==37.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1.2
django-heroku==0.3.1
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==0.8.3
fastjsonschema==2.16.2
Flask==2.2.2
fonttools==4.38.0
gunicorn==20.1.0
h11==0.14.0
idna==3.3
install==1.3.5
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.1
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.3.5
jupyter-console==6.4.3
jupyter_core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.1.2
jupyterlab_server==2.15.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==0.8.4
mysql-connector-python==8.0.31
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.5.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.3.1
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf==3.20.1
psutil==5.9.0
psycopg2==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
s3transfer==0.6.0
selenium==4.5.0
Send2Trash==1.8.0
setuptools==63.4.1
sip==6.6.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.3
stack-data==0.2.0
static3==0.7.0
terminado==0.13.1
testpath==0.6.0
toml==0.10.2
tornado==6.2
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.11
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
wheel==0.37.1
whitenoise==6.2.0
widgetsnbextension==3.5.2
wsproto==1.2.0
